Add sanity_check for removing filepath and migrate htpasswd
This commit ensures that the filename and it's variations
are removed from any HTPasswdPasswordIdentityProvider
if openshift_master_manage_htpasswd == False.  If
openshift_master_manage_htpasswd is False, we do not manage
the htpasswd file and it may be managed by a user's own automated
process outside of openshift-ansible.  We need to fail and inform
user that the file needs to be moved so they can update their
processes.

In cases where we manage an htpasswd file, we will
hardcode 'filename' to '/etc/origin/master/htpasswd'
for all new installs and upgrades.

Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=1565447